refactor(utils): clean up
