+++ packages/utils/src/asserts.d.ts
@@ -1,37 +1,3 @@
-import {ClassDescriptor, ExtendedPropertyDescriptor} from '@corpuscule/typings';
-
-export interface AssertOptions {
-  readonly correct?: boolean;
-  readonly customMessage?: string;
-}
-
-export enum Kind {
-  Accessor,
-  Class,
-  Field,
-  Getter,
-  Method,
-  Setter,
-}
-
-export enum Placement {
-  Own,
-  Prototype,
-  Static,
-}
-
-export const assertKind: (
-  decoratorName: string,
-  types: number,
-  descriptor: ExtendedPropertyDescriptor | ClassDescriptor,
-) => void;
-
-export const assertPlacement: (
-  decoratorName: string,
-  types: number,
-  descriptor: ExtendedPropertyDescriptor,
-) => void;
-
 export const assertRequiredProperty: {
   (decoratorName: string, markerName: string, propertyName: string, property: unknown): void;
   (decoratorName: string, markerName: string, property: unknown): void;
