--- packages/utils/src/descriptorsNew.d.ts
@@ -1,4 +1,0 @@
-export const makeAccessor: <T>(
-  target: unknown,
-  descriptor: PropertyDescriptor,
-) => {readonly get: () => T; readonly set: (value: T) => void};
