--- packages/utils/src/getSupersNew.d.ts
@@ -1,7 +1,0 @@
-declare const getSupers: <N extends PropertyKey>(
-  target: unknown,
-  names: ReadonlyArray<N>,
-  fallbacks?: Partial<Record<N, Function>>,
-) => Record<N, Function>;
-
-export default getSupers;
