--- packages/utils/src/getSupersNew.js
@@ -1,11 +1,0 @@
-/* eslint-disable no-empty-function */
-const noop = () => {};
-
-const getSupers = (target, names, fallbacks = {}) =>
-  names.reduce((supers, name) => {
-    supers[name] = target.prototype[name] || fallbacks[name] || noop;
-
-    return supers;
-  }, {});
-
-export default getSupers;
