--- packages/utils/src/lifecycleDescriptors.js
@@ -1,20 +1,0 @@
-import {field, method as mmethod} from './descriptors';
-
-export const method = ({key, method: value}, supers, getConstructor) => {
-  const $$key = Symbol();
-
-  return [
-    mmethod({
-      key,
-      method(...args) {
-        this[$$key](...args);
-      },
-    }),
-    field({
-      initializer() {
-        return this.constructor === getConstructor() ? value : supers[key];
-      },
-      key: $$key,
-    }),
-  ];
-};
