+++ scripts/project.js
@@ -1,13 +1,8 @@
 const utils = [
   'asserts',
   'compose',
-  'define',
-  'descriptors',
-  'descriptorsNew',
   'getSupers',
-  'getSupersNew',
-  'initUtils',
-  'lifecycleDescriptors',
+  'makeAccessor',
   'propertyUtils',
   'scheduler',
   'setters',
